Update js for International Users
After sharing project with users via the Gitter chatroom, it was discovered that the temperature toggle button wasn't functioning properly for users outside the US. This was due to an error in the 'temp' variable assignment. In addition, the ipinfo API wasn't functioning with international users due to the use of the 'zipcode' parameter. The ipinfo was then altered to use the 'city' parameter instead which was verified via Gitter international users.